+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 0.13.0 (TBD)
 
 - Added integration to import mint private notes directly to the wallet ([#189](https://github.com/0xMiden/miden-faucet/pull/189)).
+- [BREAKING] Replaced the `api-bind-url` param for `api-bind-port` ([#156](https://github.com/0xMiden/miden-faucet/pull/156)).
+- [BREAKING] Replaced the `frontend-url` param for `frontend-bind-port` ([#156](https://github.com/0xMiden/miden-faucet/pull/156)).
+- [BREAKING] Added `no-frontend` param to optionally disable the frontend server ([#156](https://github.com/0xMiden/miden-faucet/pull/156)).
 
 ## 0.12.4 (2025-12-04)
 
@@ -14,6 +17,7 @@
 
 - Fixed challenge cache lock poisoned bug by removing validation on `challenges_timestamps` cleanup ([#165](https://github.com/0xMiden/miden-faucet/pull/165)).
 - Added an error display for failed metadata requests ([#166](https://github.com/0xMiden/miden-faucet/pull/166)).
+- Improved rate limiter by tracking challenges by submission timestamp ([#167](https://github.com/0xMiden/miden-faucet/pull/167)).
 
 ## 0.12.2 (2025-11-12)